refact/front/DriverButton.jsx: moves to src/components/DriverButton.jsx
