+++ frontend/src/pages/Home.jsx
@@ -1,7 +1,7 @@
 import { useSelector } from 'react-redux';
 import sortTripsByDate from '../store/utils/sortTripsByDate';
 import CardTrip from '../components/CardTrip';
-import DriverButton from '../components/DriverButton';
+import DriverButton from '../components/buttons/DriverButton';
 import mainProfilePicture from '../assets/main-profilepic.png';
 import auto from '../assets/auto.png';
 import DateDisplay from '../components/DateDisplay';
